requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.3.0
Django==2.1.1
django-cors-middleware==1.3.1
django-extensions==2.1.2
djangorestframework==3.8.2
flake8==3.5.0
gunicorn==19.9.0
ipdb==0.11
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
mccabe==0.6.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.4.2
python-dotenv==0.9.1
pytz==2018.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
whitenoise==4.1
